resolve some of requsted change Relates #14
