+++ server/utils/index.ts
@@ -1,6 +1,7 @@
 import { verifyToken, signToken } from './jwt';
 import { hashPassword, comparePassword } from './bcrypt';
+import mailSender from './sendEmailVerification/sendEmail';
 
 export {
-  verifyToken, signToken, hashPassword, comparePassword,
+  verifyToken, signToken, hashPassword, comparePassword, mailSender,
 };
